+++ assets/components/ajaxupload/connector.php
@@ -1,33 +1,37 @@
 <?php
 /**
- * AjaxUpload Connector
+ * AjaxUpload connector
  *
  * @package ajaxupload
  * @subpackage connector
  *
  * @var modX $modx
  */
-/* Allow anonymous users */
-define('MODX_REQP', false);
+
+// Allow anonymous users for web/upload processor
+if (isset($_REQUEST['action']) && $_REQUEST['action'] == 'web/upload') {
+    define('MODX_REQP', false);
+}
 
 require_once dirname(dirname(dirname(dirname(__FILE__)))) . '/config.core.php';
 require_once MODX_CORE_PATH . 'config/' . MODX_CONFIG_KEY . '.inc.php';
 require_once MODX_CONNECTORS_PATH . 'index.php';
 
 $corePath = $modx->getOption('ajaxupload.core_path', null, $modx->getOption('core_path') . 'components/ajaxupload/');
-
 /** @var AjaxUpload $ajaxupload */
 $ajaxupload = $modx->getService('ajaxupload', 'AjaxUpload', $corePath . 'model/ajaxupload/', array(
     'core_path' => $corePath
 ));
 
-if ($_REQUEST['action'] == 'web/upload') {
+// Set HTTP_MODAUTH for web processors
+if (defined('MODX_REQP') && MODX_REQP === false) {
     $_SERVER['HTTP_MODAUTH'] = $modx->user->getUserToken($modx->context->get('key'));
 }
 
+$processorsPath = $ajaxupload->getOption('processorsPath');
+
 // Handle request
-$path = $modx->getOption('processorsPath', $ajaxupload->config, $corePath . 'processors/');
 $modx->request->handleRequest(array(
-    'processors_path' => $path,
+    'processors_path' => $processorsPath,
     'location' => ''
 ));