+++ core/components/ajaxupload/docs/changelog.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.7] - 2019-05-16
+### Changed
+- Class based upload processor
+- Update russian lexicon
+- Fix ajaxuploadAllowOverwrite option
+
 ## [1.5.6] - 2019-02-22
 ### Changed
 - Fix uploading a file with the same name